Move SAI iterator files into package iterator
This change automatically cleaned up unused imports in affected files.

Fixed few typos in affected files.
