+++ src/java/org/apache/cassandra/index/sai/iterators/RangeIntersectionIterator.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.cassandra.index.sai.utils;
+package org.apache.cassandra.index.sai.iterators;
 
 import java.io.IOException;
 import java.lang.invoke.MethodHandles;
@@ -28,6 +28,7 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.apache.cassandra.index.sai.utils.PrimaryKey;
 import org.apache.cassandra.io.util.FileUtils;
 
 /**
@@ -69,7 +70,7 @@ public class RangeIntersectionIterator extends RangeIterator
         // It can become null when we reach the end of the iterator.
         PrimaryKey highestKey = ranges.get(0).hasNext() ? ranges.get(0).next() : null;
         // Index of the range iterator that has advanced beyond the others
-        int alreadyAvanced = 0;
+        int alreadyAdvanced = 0;
         rangeStats[0]++;
 
         outer:
@@ -79,7 +80,7 @@ public class RangeIntersectionIterator extends RangeIterator
             // Once this inner loop finishes normally, all iterators are guaranteed to be at the same value.
             for (int index = 0; index < ranges.size(); index++)
             {
-                if (index != alreadyAvanced)
+                if (index != alreadyAdvanced)
                 {
                     RangeIterator range = ranges.get(index);
                     PrimaryKey nextKey = nextOrNull(range, highestKey);
@@ -90,7 +91,7 @@ public class RangeIntersectionIterator extends RangeIterator
                         // We jumped over the highest key seen so far, so make it the new highest key.
                         highestKey = nextKey;
                         // Remember this iterator to avoid advancing it again, because it is already at the highest key
-                        alreadyAvanced = index;
+                        alreadyAdvanced = index;
                         // This iterator jumped over, so the other iterators are lagging behind now,
                         // including the ones already advanced in the earlier cycles of the inner loop.
                         // Therefore, restart the inner loop in order to advance
