+++ test/unit/org/apache/cassandra/index/sai/iterators/AbstractRangeIteratorTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.cassandra.index.sai.utils;
+package org.apache.cassandra.index.sai.iterators;
 
 import java.util.Arrays;
 import java.util.Set;
@@ -23,6 +23,8 @@ import java.util.stream.Collectors;
 
 import org.junit.Assert;
 
+import org.apache.cassandra.index.sai.utils.PrimaryKey;
+import org.apache.cassandra.index.sai.utils.SaiRandomizedTest;
 import org.apache.cassandra.utils.Pair;
 
 public class AbstractRangeIteratorTest extends SaiRandomizedTest
