+++ test/unit/org/apache/cassandra/index/sai/iterators/RangeConcatIteratorTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.cassandra.index.sai.utils;
+package org.apache.cassandra.index.sai.iterators;
 
 import java.util.ArrayList;
 import java.util.function.Supplier;
@@ -25,7 +25,7 @@ import org.junit.Test;
 
 import org.apache.cassandra.utils.Pair;
 
-import static org.apache.cassandra.index.sai.utils.LongIterator.convert;
+import static org.apache.cassandra.index.sai.iterators.LongIterator.convert;
 
 public class RangeConcatIteratorTest extends AbstractRangeIteratorTest
 {
